restructure batching logic for efficiency restructure batching logic for efficient querying
